@@ -8,49 +8,58 @@
     "npm": ">=6"
   },
   "dependencies": {
-    "@emotion/core": "^10.0.27",
+    "@emotion/core": "^10.0.28",
     "@emotion/styled": "^10.0.27",
-    "@reach/dialog": "^0.8.0",
-    "@reach/menu-button": "^0.8.0",
-    "@reach/router": "^1.2.1",
-    "@reach/tabs": "^0.8.0",
-    "@reach/tooltip": "^0.8.0",
-    "@reach/visually-hidden": "^0.8.0",
-    "@testing-library/jest-dom": "^5.0.2",
-    "@testing-library/react": "^9.4.0",
+    "@reach/dialog": "^0.9.1",
+    "@reach/menu-button": "^0.9.1",
+    "@reach/tabs": "^0.9.1",
+    "@reach/tooltip": "^0.9.1",
+    "@reach/visually-hidden": "^0.9.0",
+    "@testing-library/jest-dom": "^5.1.1",
+    "@testing-library/react": "^10.0.1",
+    "@testing-library/user-event": "^10.0.0",
+    "autoprefixer": "^9.7.4",
     "bootstrap": "^4.4.1",
-    "debounce-fn": "^3.0.1",
+    "debounce-fn": "^4.0.0",
     "faker": "^4.1.0",
+    "history": "^5.0.0-beta.5",
     "match-sorter": "^4.0.2",
+    "msw": "^0.10.0",
+    "node-match-path": "^0.3.1",
     "prop-types": "^15.7.2",
     "querystringify": "^2.1.1",
-    "react": "16.12.0",
-    "react-async": "^10.0.0",
-    "react-dom": "16.12.0",
+    "react": "16.13.1",
+    "react-dom": "16.13.1",
+    "react-error-boundary": "^1.2.5",
     "react-icons": "^3.9.0",
-    "react-scripts": "^3.3.1",
-    "serve": "^11.3.0",
-    "stop-runaway-react-effects": "^1.2.1"
+    "react-query": "^1.0.34",
+    "react-query-devtools": "^1.0.12",
+    "react-router": "^6.0.0-alpha.2",
+    "react-router-dom": "^6.0.0-alpha.2",
+    "react-scripts": "^3.4.1",
+    "stop-runaway-react-effects": "^2.0.0"
   },
   "devDependencies": {
-    "@testing-library/cypress": "^5.1.2",
-    "cross-env": "^7.0.0",
-    "cypress": "^3.8.3",
-    "cypress-hmr-restarter": "^1.0.1",
-    "eslint-plugin-cypress": "^2.8.1",
-    "husky": "^4.2.1",
+    "@testing-library/cypress": "^6.0.0",
+    "cross-env": "^7.0.2",
+    "cypress": "^4.2.0",
+    "cypress-hmr-restarter": "^1.0.2",
+    "eslint-plugin-cypress": "^2.10.3",
+    "husky": "^4.2.3",
     "is-ci-cli": "^2.0.0",
+    "jest-environment-jsdom-sixteen": "^1.0.3",
     "npm-run-all": "^4.1.5",
-    "prettier": "^1.19.1",
-    "start-server-and-test": "^1.10.8",
+    "prettier": "^2.0.1",
+    "serve": "^11.3.0",
+    "start-server-and-test": "^1.10.11",
     "test-data-bot": "^0.8.0"
   },
   "scripts": {
     "start": "react-scripts start",
     "build": "react-scripts build",
-    "test": "react-scripts test",
+    "test": "react-scripts test --env=jest-environment-jsdom-sixteen",
     "test:coverage": "npm run test -- --watchAll false --coverage",
-    "test:debug": "node --inspect-brk ./node_modules/react-scripts/scripts/test.js --runInBand",
+    "test:debug": "node --inspect-brk ./node_modules/react-scripts/scripts/test.js --env=jest-environment-jsdom-sixteen --watch --runInBand",
     "cy:run": "cypress run",
     "cy:open": "cypress open",
     "test:e2e": "is-ci \"test:e2e:run\" \"test:e2e:dev\"",
@@ -70,6 +79,14 @@
     "hooks": {
       "pre-commit": "npm run validate"
     }
+  },
+  "jest": {
+    "collectCoverageFrom": [
+      "src/**/*.js",
+      "!<rootDir>/node_modules/**/*",
+      "!<rootDir>/src/test/**/*",
+      "!<rootDir>/src/dev-tools/**/*"
+    ]
   },
   "browserslist": {
     "development": [
