@@ -1,98 +1,143 @@
 /** @jsx jsx */
+/** @jsxFrag React.Fragment */
 import {jsx} from '@emotion/core'
 
-import styled from '@emotion/styled'
-import {Router, Link, Redirect} from '@reach/router'
+import React from 'react'
+import {
+  Routes,
+  Route,
+  Link as RouterLink,
+  useNavigate,
+  useMatch,
+} from 'react-router-dom'
+import ErrorBoundary from 'react-error-boundary'
 import * as mq from './styles/media-queries'
 import * as colors from './styles/colors'
-import {ListItemProvider} from './context/list-item-context'
 import {useAuth} from './context/auth-context'
-import {useUser} from './context/user-context'
 import ReadingListScreen from './screens/list'
 import FinishedBooksScreen from './screens/finished'
 import DiscoverBooksScreen from './screens/discover'
 import BookScreen from './screens/book'
 import NotFound from './screens/not-found'
 
-function AuthenticatedApp() {
-  const user = useUser()
-  const {logout} = useAuth()
+function ErrorFallback({error}) {
   return (
-    <div
-      css={{
-        margin: '0 auto',
-        padding: '2em 0',
-        maxWidth: '840px',
-        width: '100%',
-        display: 'grid',
-        gridGap: '1em',
-        gridTemplateColumns: '3fr 9fr',
-        [mq.small]: {
-          gridTemplateColumns: '1fr',
-          gridTemplateRows: 'auto',
-          width: '100%',
-          padding: '2em 1em',
-        },
-      }}
-    >
-      <div>
-        <div
-          css={{
-            display: 'flex',
-            alignItems: 'center',
-            position: 'absolute',
-            top: '10px',
-            right: '10px',
-          }}
-        >
-          {user.username}
-          <button
-            onClick={logout}
-            css={{
-              marginLeft: '10px',
-            }}
-          >
-            Logout
-          </button>
-        </div>
-        <Nav />
-      </div>
-      <main css={{width: '100%'}}>
-        <Routes />
-      </main>
-      <footer />
+    <div role="alert" css={{color: colors.danger, fontSize: '0.7em'}}>
+      <span>There was an error:</span>{' '}
+      <pre
+        css={{
+          display: 'inline-block',
+          overflow: 'scroll',
+          margin: '0',
+          marginBottom: -5,
+        }}
+      >
+        {error.message}
+      </pre>
     </div>
   )
 }
 
-const NavLink = styled(Link)({
-  display: 'block',
-  padding: '8px 15px 8px 10px',
-  margin: '5px 0',
-  width: '100%',
-  height: '100%',
-  color: colors.text,
-  borderRadius: '2px',
-  borderLeft: '5px solid transparent',
-  ':hover': {
-    color: colors.indigo,
-    textDecoration: 'none',
-    background: colors.gray10,
-  },
-})
+function AuthenticatedApp() {
+  const {user, logout} = useAuth()
+  return (
+    <>
+      <div
+        css={{
+          display: 'flex',
+          alignItems: 'center',
+          position: 'absolute',
+          top: '10px',
+          right: '10px',
+        }}
+      >
+        {user.username}
+        <button
+          css={{
+            marginLeft: '10px',
+          }}
+          onClick={logout}
+        >
+          Logout
+        </button>
+      </div>
+      <div
+        css={{
+          margin: '0 auto',
+          padding: '4em 2em',
+          maxWidth: '840px',
+          width: '100%',
+          display: 'grid',
+          gridGap: '1em',
+          gridTemplateColumns: '3fr 9fr',
+          [mq.small]: {
+            gridTemplateColumns: '1fr',
+            gridTemplateRows: 'auto',
+            width: '100%',
+          },
+        }}
+      >
+        <div css={{position: 'relative'}}>
+          <Nav />
+        </div>
+        <main css={{width: '100%'}}>
+          <ErrorBoundary FallbackComponent={ErrorFallback}>
+            <AppRoutes />
+          </ErrorBoundary>
+        </main>
+        <footer />
+      </div>
+    </>
+  )
+}
+
+function NavLink(props) {
+  const match = useMatch(props.to)
+  return (
+    <RouterLink
+      css={[
+        {
+          display: 'block',
+          padding: '8px 15px 8px 10px',
+          margin: '5px 0',
+          width: '100%',
+          height: '100%',
+          color: colors.text,
+          borderRadius: '2px',
+          borderLeft: '5px solid transparent',
+          ':hover': {
+            color: colors.indigo,
+            textDecoration: 'none',
+            background: colors.gray10,
+          },
+        },
+        match
+          ? {
+              borderLeft: `5px solid ${colors.indigo}`,
+              background: colors.gray10,
+              ':hover': {
+                background: colors.gray10,
+              },
+            }
+          : null,
+      ]}
+      {...props}
+    />
+  )
+}
 
 function Nav(params) {
   return (
     <nav
       css={{
         position: 'sticky',
-        top: '2em',
-        padding: '3em 1.5em',
+        top: '4px',
+        padding: '1em 1.5em',
         border: `1px solid ${colors.gray10}`,
         borderRadius: '3px',
         [mq.small]: {
-          padding: '0.5em 1em',
           position: 'static',
+          top: 'auto',
         },
       }}
     >
@@ -100,13 +145,6 @@ function Nav(params) {
         css={{
           listStyle: 'none',
           padding: '0',
-          '& [aria-current="page"]': {
-            borderLeft: `5px solid ${colors.indigo}`,
-            background: colors.gray10,
-            ':hover': {
-              background: colors.gray10,
-            },
-          },
         }}
       >
         <li>
@@ -124,21 +162,23 @@ function Nav(params) {
 }
 
 function RedirectHome() {
-  return <Redirect noThrow to="/list" />
+  const navigate = useNavigate()
+  React.useEffect(() => {
+    navigate('/list', {replace: true})
+  }, [navigate])
+  return null
 }
 
-function Routes() {
+function AppRoutes() {
   return (
-    <ListItemProvider>
-      <Router>
-        <RedirectHome path="/" />
-        <ReadingListScreen path="/list" />
-        <FinishedBooksScreen path="/finished" />
-        <DiscoverBooksScreen path="/discover" />
-        <BookScreen path="/book/:bookId" />
-        <NotFound default />
-      </Router>
-    </ListItemProvider>
+    <Routes>
+      <Route path="/" element={<RedirectHome />} />
+      <Route path="/list" element={<ReadingListScreen />} />
+      <Route path="/finished" element={<FinishedBooksScreen />} />
+      <Route path="/discover" element={<DiscoverBooksScreen />} />
+      <Route path="/book/:bookId" element={<BookScreen />} />
+      <Route path="*" element={<NotFound />} />
+    </Routes>
   )
 }
 
