@@ -1,18 +1,14 @@
 import React from 'react'
-import {useUser} from './context/user-context'
+import {useAuth} from './context/auth-context'
 import {FullPageSpinner} from './components/lib'
 
-const loadAuthenticatedApp = () => import('./authenticated-app')
-const AuthenticatedApp = React.lazy(loadAuthenticatedApp)
+const AuthenticatedApp = React.lazy(() =>
+  import(/* webpackPrefetch: true */ './authenticated-app'),
+)
 const UnauthenticatedApp = React.lazy(() => import('./unauthenticated-app'))
 
 function App() {
-  const user = useUser()
-  // pre-load the authenticated side in the background while the user's
-  // filling out the login form.
-  React.useEffect(() => {
-    loadAuthenticatedApp()
-  }, [])
+  const {user} = useAuth()
   return (
     <React.Suspense fallback={<FullPageSpinner />}>
       {user ? <AuthenticatedApp /> : <UnauthenticatedApp />}
