@@ -1,5 +1,5 @@
-import './hacks'
-import 'bootstrap/dist/css/bootstrap-reboot.css'
+import 'stop-runaway-react-effects/hijack'
+import './test/server'
 import '@reach/dialog/styles.css'
 import '@reach/menu-button/styles.css'
 import '@reach/tabs/styles.css'
