@@ -1,54 +1,77 @@
 /** @jsx jsx */
+/** @jsxFrag React.Fragment */
 import {jsx} from '@emotion/core'
 
 import React from 'react'
-import {useAsync} from 'react-async'
 import {bootstrapAppData} from '../utils/bootstrap'
 import * as authClient from '../utils/auth-client'
+import useAsync from '../utils/use-async'
+import * as colors from '../styles/colors'
 import {FullPageSpinner} from '../components/lib'
 
 const AuthContext = React.createContext()
+AuthContext.displayName = 'AuthContext'
+
+const appDataPromise = bootstrapAppData()
 
 function AuthProvider(props) {
-  const [firstAttemptFinished, setFirstAttemptFinished] = React.useState(false)
-  const {
-    data = {user: null, listItems: []},
-    error,
-    isRejected,
-    isPending,
-    isSettled,
-    reload,
-  } = useAsync({
-    promiseFn: bootstrapAppData,
-  })
+  const {data, status, error, isLoading, isError, isSuccess, run} = useAsync()
+
+  const runBootstrap = React.useCallback(() => run(bootstrapAppData()), [run])
 
   React.useLayoutEffect(() => {
-    if (isSettled) {
-      setFirstAttemptFinished(true)
-    }
-  }, [isSettled])
+    run(appDataPromise)
+  }, [run])
 
-  if (!firstAttemptFinished) {
-    if (isPending) {
-      return <FullPageSpinner />
-    }
-    if (isRejected) {
-      return (
-        <div css={{color: 'red'}}>
-          <p>Uh oh... There's a problem. Try refreshing the app.</p>
-          <pre>{error.message}</pre>
-        </div>
-      )
-    }
+  const login = React.useCallback(
+    form => authClient.login(form).then(runBootstrap),
+    [runBootstrap],
+  )
+  const register = React.useCallback(
+    form => authClient.register(form).then(runBootstrap),
+    [runBootstrap],
+  )
+  const logout = React.useCallback(() => {
+    authClient.logout()
+    run(bootstrapAppData())
+  }, [run])
+
+  const user = data?.user
+
+  const value = React.useMemo(() => ({user, login, logout, register}), [
+    login,
+    logout,
+    register,
+    user,
+  ])
+
+  if (isLoading) {
+    return <FullPageSpinner />
   }
 
-  const login = form => authClient.login(form).then(reload)
-  const register = form => authClient.register(form).then(reload)
-  const logout = () => authClient.logout().then(reload)
+  if (isError) {
+    return (
+      <div
+        css={{
+          color: colors.danger,
+          height: '100vh',
+          display: 'flex',
+          flexDirection: 'column',
+          justifyContent: 'center',
+          alignItems: 'center',
+        }}
+      >
+        <p>Uh oh... There's a problem. Try refreshing the app.</p>
+        <pre>{error.message}</pre>
+      </div>
+    )
+  }
 
-  return (
-    <AuthContext.Provider value={{data, login, logout, register}} {...props} />
-  )
+  if (isSuccess) {
+    return <AuthContext.Provider value={value} {...props} />
+  }
+
+  throw new Error(`Unhandled status: ${status}`)
 }
 
 function useAuth() {
