@@ -1,12 +1,15 @@
+import loadDevTools from './dev-tools/load'
 import './bootstrap'
 import React from 'react'
 import ReactDOM from 'react-dom'
 import App from './app'
 import AppProviders from './context'
 
-ReactDOM.render(
-  <AppProviders>
-    <App />
-  </AppProviders>,
-  document.getElementById('root'),
-)
+loadDevTools(() => {
+  ReactDOM.render(
+    <AppProviders>
+      <App />
+    </AppProviders>,
+    document.getElementById('root'),
+  )
+})
