@@ -1,5 +1,5 @@
 import React from 'react'
-import {render} from '@testing-library/react'
+import {render} from 'test/app-test-utils'
 import NotFound from '../not-found'
 
 test('renders a go home link', () => {
