@@ -3,7 +3,7 @@ import client from '../api-client'
 const url = endpoint => `${process.env.REACT_APP_API_URL}/${endpoint}`
 
 test('calls fetch at the endpoint with the arguments for GET requests', async () => {
-  window.fetch.mockResolvedValueOnce({json: () => Promise.resolve()})
+  window.fetch.mockResolvedValueOnce({ok: true, json: () => Promise.resolve()})
   await client('foo')
   expect(window.fetch).toHaveBeenCalledTimes(1)
   expect(window.fetch).toHaveBeenCalledWith(url('foo'), {
@@ -16,7 +16,7 @@ test('calls fetch at the endpoint with the arguments for GET requests', async ()
 
 test('adds auth token when a token is in localStorage', async () => {
   window.localStorage.setItem('__bookshelf_token__', 'FAKE_TOKEN')
-  window.fetch.mockResolvedValueOnce({json: () => Promise.resolve()})
+  window.fetch.mockResolvedValueOnce({ok: true, json: () => Promise.resolve()})
   await client('foo')
   expect(window.fetch).toHaveBeenCalledTimes(1)
   expect(window.fetch).toHaveBeenCalledWith(url('foo'), {
@@ -30,7 +30,7 @@ test('adds auth token when a token is in localStorage', async () => {
 })
 
 test('allows for config overrides', async () => {
-  window.fetch.mockResolvedValueOnce({json: () => Promise.resolve()})
+  window.fetch.mockResolvedValueOnce({ok: true, json: () => Promise.resolve()})
   await client('foo', {
     credentials: 'omit',
     headers: {'content-type': 'fake-type'},
