@@ -1,17 +1,25 @@
-import {getUser} from './auth-client'
-import {readForUser} from './list-items-client'
+import {queryCache} from 'react-query'
+import * as auth from './auth-client'
+import * as listItemsClient from './list-items-client'
 
 async function bootstrapAppData() {
-  const data = await getUser()
-  if (!data) {
-    return {user: null, listItems: []}
+  let appData = {user: null, listItems: []}
+
+  if (auth.isLoggedIn()) {
+    try {
+      await window.__bookshelf_serverReady
+      const [user, listItems] = await Promise.all([
+        auth.getUser().then(d => d.user),
+        listItemsClient.read().then(d => d.listItems),
+      ])
+      appData = {user, listItems}
+    } catch (error) {
+      auth.logout()
+      throw error
+    }
   }
-  const {user} = data
-  const {listItems} = await readForUser(user.id)
-  return {
-    user,
-    listItems,
-  }
+  queryCache.setQueryData('list-items', appData.listItems)
+  return appData
 }
 
 export {bootstrapAppData}
